+++ tensorflow_recommenders/layers/__init__.py
@@ -15,5 +15,6 @@
 # Lint-as: python3
 """Layers."""
 
+from tensorflow_recommenders.layers import dcn
 from tensorflow_recommenders.layers import factorized_top_k
 from tensorflow_recommenders.layers import loss