+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unlreased][unreleased]
+## [0.3.0][2020-11-18]
 
 ### Added